requirements.txt
exceptiongroup==1.2.2
greenlet==3.0.3
iniconfig==2.0.0
mysql-connector-python==9.0.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.32
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
